+++ Kanikama/Packages/net.shivaduke28.kanikama.gi/Baking/Impl/KanikamaLight.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 namespace Kanikama.GI.Baking.Impl
 {
     [RequireComponent(typeof(Light))]
-    [AddComponentMenu("Kanikama/GI/Baking/KanikamaBakeTargetLight")]
+    [AddComponentMenu("Kanikama/GI/Baking.KanikamaLight")]
     [EditorOnly]
     public sealed class KanikamaLight : BakeTarget
     {
