+++ Kanikama/Packages/net.shivaduke28.kanikama.gi/Runtime/Impl/KanikamaGIUpdater.cs
@@ -3,7 +3,7 @@ using UnityEngine;
 
 namespace Kanikama.GI.Runtime.Impl
 {
-    [AddComponentMenu("Kanikama/GI/Runtime/KanikamaGIUpdater")]
+    [AddComponentMenu("Kanikama/GI/Runtime.KanikamaGIUpdater")]
     public class KanikamaGIUpdater : MonoBehaviour
     {
         [SerializeField] KanikamaSceneDescriptor kanikamaSceneDescriptor;
