+++ Kanikama/Packages/net.shivaduke28.kanikama.gi/Runtime/Impl/KanikamaLight.cs
@@ -3,7 +3,7 @@
 namespace Kanikama.GI.Runtime.Impl
 {
     [RequireComponent(typeof(Light))]
-    [AddComponentMenu("Kanikama/GI/Runtime/KanikamaLight")]
+    [AddComponentMenu("Kanikama/GI/Runtime.KanikamaLight")]
     public sealed class KanikamaLight : LightSource
     {
         [SerializeField] new Light light;
