+++ Kanikama/Packages/net.shivaduke28.kanikama.gi/Runtime/Impl/KanikamaSceneDescriptor.cs
@@ -4,7 +4,7 @@ using UnityEngine;
 
 namespace Kanikama.GI.Runtime.Impl
 {
-    [AddComponentMenu("Kanikama/GI/Runtime/KanikamaSceneDescriptor")]
+    [AddComponentMenu("Kanikama/GI/Runtime.KanikamaSceneDescriptor")]
     public sealed class KanikamaSceneDescriptor : MonoBehaviour
     {
         [SerializeField] List<LightSource> lightSources;
